✨ JE-72-feat: 마이페이지용 애니메이션 설정
